+++ lib/ts/recipe/thirdparty/providers/github.ts
@@ -15,7 +15,6 @@
 import { TypeProvider, TypeProviderGetResponse } from "../types";
 import axios from "axios";
 import { validateTheStructureOfUserInput } from "../../../utils";
-import Recipe from "../recipe";
 import { Request } from "express";
 
 type TypeThirdPartyProviderGithubConfig = {